Move dataroot
