+++ src/sastadev/constants.py
@@ -1,0 +1,14 @@
+'''
+Module to specify constant values that are used in the application
+'''
+
+# default locations of the data to train and test sastadev
+bronzefolder = 'bronze'
+formsfolder = 'forms'
+indatafolder = 'indata'
+intreebanksfolder = 'intreebanks'
+loggingfolder = 'logging'
+outtreebanksfolder = 'outtreebanks'
+resultsfolder = 'results'
+silverfolder = 'silver'
+silverpermfolder = 'silverperm'
